requirements.txt
flask==3.0.3
werkzeug==3.0.3
gunicorn==21.2.0
python-dotenv==1.0.0
flask-wtf==1.2.1
email-validator==2.2.0
